Fix bug when the rounding adjustment is positive
Added some mandatory params when the discount doesn't be a correct type for the rounding adjustment. If you try to fix the totals using an item without type, the API return an error requesting this params:
 - price_with_tax
 - quantity
 - downloadble

I supose that is not the best solution but is a temporal solution that allow to use the php-client.